Refactor module structure
#3
