+++ pom.xml
@@ -6,7 +6,11 @@
 
     <groupId>io.wildlabs</groupId>
     <artifactId>filemonster</artifactId>
+    <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
+    <modules>
+        <module>filemonster-core</module>
+    </modules>
 
     <properties>
         <maven.compiler.source>1.6</maven.compiler.source>
